feat(API/signup): email on successful signup
